backend/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysql-connector-python==8.0.12
protobuf==3.6.1
six==1.11.0
Werkzeug==0.14.1
